feat,fix,ui/detail,transaction: feat,ui: detail 판매중,거래완료 기능 및 UI 추가 fix: detail,transaction 컴포넌트 분리
